Refactor: extract snapshot_io for minimal container deps
Move fetch_snapshot_to_path to props/core/db/snapshot_io.py to reduce
the dependency footprint for in-container agent helpers. This removes
the CLI dependencies (pygit2, typer, yaml, sync module) from the
critic container's transitive deps.

- Create snapshot_io.py with just fetch_snapshot_to_path
- Update agent_helpers.py to import from snapshot_io
- Update cmd_snapshot.py to import from snapshot_io
- Remove stale :agent_registry dep from agent_helpers BUILD
